Back buttons go to previous page
Made the back buttons go to home as that is the actual previous page in the flow. It's unintuitive to the user if the back button takes them to a different page. (Let me know if this is not how you want it done though and we can change it back).